Rename PageItem to Contribution and add short version
We'd like to be able to copy in two formats. For example,
For a PR named "fix linting issues #123", we should offer
two quick links:
1) fix linting issues #123
2) #123
Previously copied list should also remember which format is
copied. Currently it doesn't do that, it just lists it
in the longer format. That should be fixed later for 1.0